New request-swype-code cgi request
Swype-code request is split into two phases. In the phase one the user
declares that he needs a swype-code. The request is passed to the server
as a transaction of requestSwypeCode method invocation. Transaction hash
is returned. In the phase two the user receives the actual swype-code
which is calculated from both transaction hash and block hash that
includes this transaction. The swype-code can be generated only after
the confirmation of the transaction.