Fast mode swype-code request
In this less secure mode the code is generated from user address and
hash of the block near the top of the blockchain. Swype code generation
is not recoded in blockchain in any way. Client just receives the swype
code and the reference block hash which is used instead of the hash of
requestSwypeCode transaction.